+++ src/config.ts
@@ -139,6 +139,7 @@ export const config = {
   maxDepthOfRepoSearch: (): number => getConfig("maxDepthOfRepoSearch", 0),
   muteCommitsNotAncestorsOfHead: (): boolean =>
     getConfig("repository.muteCommitsNotAncestorsOfHead", false),
+  muteMergeCommits: (): boolean => getConfig("repository.muteMergeCommits", false),
   onlyFollowFirstParent: (): boolean => getConfig("repository.onlyFollowFirstParent", false),
   showCurrentBranchByDefault: (): boolean => getConfig("showCurrentBranchByDefault", false),
   showRemoteBranches: (): boolean => getConfig("repository.showRemoteBranches", true),